Refactor ingredients input into reusable component with enhanced functionality
