+++ components/dialogs/add-recipe-dialog.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { LexicalEditor } from "@/components/rich-text/lexical-editor";
-import { Badge } from "@/components/ui/badge";
 import { BaseDialog } from "@/components/ui/base-dialog";
 import { Button } from "@/components/ui/button";
 import {
@@ -11,38 +10,17 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Checkbox } from "@/components/ui/checkbox";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
+import { IngredientsInput } from "@/components/ui/ingredients-input";
 import { useTranslations } from "@/hooks/use-translations";
-import { DEFAULT_UNIT, UNIT_OPTIONS } from "@/lib/constants/units";
+import { DEFAULT_UNIT } from "@/lib/constants/units";
 import { trimObjectStrings } from "@/lib/utils/form-utils";
-// Removed drag-and-drop imports
-import { ErrorMessage, Field, FieldArray, Form, Formik } from "formik";
+import { ErrorMessage, Field, Form, Formik } from "formik";
 import {
   BookOpen,
   CheckCircle2,
   ChefHat,
-  ChevronDown,
-  DollarSign,
-  GripVertical,
-  Package,
-  Plus,
-  Trash2,
-  X,
 } from "lucide-react";
 import type { ClipboardEvent } from "react";
 import { useCallback, useEffect, useMemo, useState } from "react";
@@ -103,8 +81,6 @@ export function AddRecipeDialog({
   const [isLoadingRecipe, setIsLoadingRecipe] = useState(false);
   const [fetchedRecipe, setFetchedRecipe] = useState<any>(null);
   const [selectedIds, setSelectedIds] = useState<Set<string>>(new Set());
-
-  // Removed drag-and-drop state and sensors
 
   // Stable ID generation to prevent unnecessary re-renders
   const generateStableId = useCallback(() => {
@@ -378,19 +354,7 @@ export function AddRecipeDialog({
     onOpenChange(false);
   };
 
-  // Calculate total cost from ingredient groups
-  const calculateTotalCost = (ingredientGroups: IngredientGroupFormValue[]) => {
-    return ingredientGroups.reduce((total, group) => {
-      return (
-        total +
-        group.ingredients.reduce((groupTotal, ingredient) => {
-          const quantity = parseFloat(ingredient.quantity) || 0;
-          const costPerUnit = parseFloat(ingredient.costPerUnit || "0") || 0;
-          return groupTotal + quantity * costPerUnit;
-        }, 0)
-      );
-    }, 0);
-  };
+
 
   return (
     <BaseDialog
@@ -422,67 +386,7 @@ export function AddRecipeDialog({
           // Ensure stable localId per ingredient for selection & DnD
           // We'll handle this in the render function instead of useEffect to avoid form interference
 
-          const toggleRowSelected = (localId: string, checked: boolean) => {
-            setSelectedIds((prev) => {
-              const next = new Set(prev);
-              if (checked) next.add(localId);
-              else next.delete(localId);
-              return next;
-            });
-          };
 
-          const setGroupSelection = (groupIndex: number, checked: boolean) => {
-            const ids = values.ingredientGroups[groupIndex].ingredients.map(
-              (ing: any) => ing.localId
-            );
-            setSelectedIds((prev) => {
-              const next = new Set(prev);
-              if (checked) ids.forEach((id: string) => next.add(id));
-              else ids.forEach((id: string) => next.delete(id));
-              return next;
-            });
-          };
-
-          const clearSelection = () => setSelectedIds(new Set());
-
-          const moveSelectedIngredients = (
-            destinationGroupIndex: number,
-            position: "end" | "start" = "end"
-          ) => {
-            if (selectedIds.size === 0) return;
-            const nextGroups = values.ingredientGroups.map((g) => ({
-              ...g,
-              ingredients: [...g.ingredients],
-            }));
-            const moved: any[] = [];
-            nextGroups.forEach((group) => {
-              const keep: any[] = [];
-              for (const ing of group.ingredients as any[]) {
-                if (selectedIds.has(ing.localId)) moved.push(ing);
-                else keep.push(ing);
-              }
-              group.ingredients = keep.length
-                ? keep
-                : [
-                    {
-                      name: "",
-                      quantity: "",
-                      unit: DEFAULT_UNIT,
-                      costPerUnit: "",
-                      localId: generateStableId(),
-                    },
-                  ];
-            });
-            const dest = nextGroups[destinationGroupIndex];
-            dest.ingredients =
-              position === "start"
-                ? [...moved, ...dest.ingredients]
-                : [...dest.ingredients, ...moved];
-            setFieldValue("ingredientGroups", nextGroups, false);
-            clearSelection();
-          };
-
-          // Removed drag-and-drop helpers and handlers
           // Show loading state while fetching recipe details
           if (isEditMode && recipeId && (isLoadingRecipe || !fetchedRecipe)) {
             return (
@@ -683,304 +587,19 @@ export function AddRecipeDialog({
               </Card>
 
               {/* Ingredient Groups Section */}
-              <FieldArray name="ingredientGroups">
-                {({ remove: removeGroup, push: pushGroup }) => (
-                  <Card>
-                    <CardHeader className="pb-4">
-                      <div className="flex items-center justify-between">
-                        <div>
-                          <CardTitle className="text-lg flex items-center gap-2">
-                            <Package className="w-5 h-5 text-primary" />
-                            Ingredient Groups
-                          </CardTitle>
-                          <CardDescription>
-                            Organize ingredients into logical groups like
-                            "Dough", "Filling", "Sauce"
-                          </CardDescription>
-                        </div>
-                        <div className="flex items-center gap-2">
-                          <Badge variant="outline" className="text-xs">
-                            {values.ingredientGroups.reduce(
-                              (total, group) =>
-                                total + group.ingredients.length,
-                              0
-                            )}{" "}
-                            ingredients
-                          </Badge>
-                          {selectedIds.size > 0 && (
-                            <div className="flex items-center gap-2">
-                              <span className="text-xs text-muted-foreground">
-                                {selectedIds.size} selected
-                              </span>
-                              <DropdownMenu>
-                                <DropdownMenuTrigger asChild>
-                                  <Button
-                                    type="button"
-                                    size="sm"
-                                    variant="outline"
-                                    className="text-xs px-3 py-1 h-7"
-                                  >
-                                    Move To <ChevronDown className="w-3 h-3 ml-1" />
-                                  </Button>
-                                </DropdownMenuTrigger>
-                                <DropdownMenuContent align="end" className="w-48">
-                                  {values.ingredientGroups.map((g, idx) => (
-                                    <DropdownMenuItem
-                                      key={idx}
-                                      onClick={() => {
-                                        // Move ingredients immediately when group is selected
-                                        moveSelectedIngredients(idx);
-                                      }}
-                                      className="cursor-pointer"
-                                    >
-                                      {g.name || `Group ${idx + 1}`}
-                                    </DropdownMenuItem>
-                                  ))}
-                                </DropdownMenuContent>
-                              </DropdownMenu>
-                              <Button
-                                type="button"
-                                variant="ghost"
-                                size="sm"
-                                onClick={clearSelection}
-                              >
-                                Clear
-                              </Button>
-                            </div>
-                          )}
-                          <Button
-                            type="button"
-                            variant="outline"
-                            size="sm"
-                            onClick={() => {
-                              pushGroup({
-                                name: "",
-                                sortOrder: values.ingredientGroups.length,
-                                ingredients: [
-                                  {
-                                    name: "",
-                                    quantity: "",
-                                    unit: DEFAULT_UNIT,
-                                    costPerUnit: "",
-                                    localId: generateStableId(),
-                                  },
-                                ],
-                              });
-                            }}
-                            className="flex items-center gap-2"
-                          >
-                            <Plus className="w-3 h-3" />
-                            Add Group
-                          </Button>
-                        </div>
-                      </div>
-                    </CardHeader>
-                    <CardContent className="space-y-6">
-                      {values.ingredientGroups.map((group, groupIndex) => (
-                        <div key={groupIndex} className="border border-border rounded-lg p-4 bg-card/30">
-                          {/* Group Header */}
-                          <div className="flex items-center justify-between mb-4">
-                            <div className="flex items-center gap-3 flex-1">
-                              <Checkbox
-                                checked={
-                                  group.ingredients.length > 0 &&
-                                  group.ingredients.every(
-                                    (ing: any) => selectedIds.has(ing.localId)
-                                  )
-                                }
-                                onCheckedChange={(checked) =>
-                                  setGroupSelection(groupIndex, Boolean(checked))
-                                }
-                                className="mr-1"
-                              />
-                              <GripVertical className="w-4 h-4 text-muted-foreground" />
-                              <div className="flex-1">
-                                <Field
-                                  as={Input}
-                                  name={`ingredientGroups[${groupIndex}].name`}
-                                  placeholder="Group name (e.g., Dough, Filling, Sauce)"
-                                  className="font-medium"
-                                />
-                                <ErrorMessage
-                                  name={`ingredientGroups[${groupIndex}].name`}
-                                  component="p"
-                                  className="text-destructive text-xs mt-1"
-                                />
-                              </div>
-                            </div>
-                            <div className="flex items-center gap-2">
-                              <Badge variant="secondary" className="text-xs">
-                                {group.ingredients.length} items
-                              </Badge>
-                              {values.ingredientGroups.length > 1 && (
-                                <Button
-                                  type="button"
-                                  variant="ghost"
-                                  size="sm"
-                                  onClick={() => removeGroup(groupIndex)}
-                                  className="text-destructive hover:bg-destructive/10"
-                                >
-                                  <Trash2 className="w-4 h-4" />
-                                </Button>
-                              )}
-                            </div>
-                          </div>
-
-                          {/* Group Ingredients */}
-                          <FieldArray name={`ingredientGroups[${groupIndex}].ingredients`}>
-                            {({ remove: removeIngredient, push: pushIngredient }) => (
-                              <div className="space-y-3" onPaste={handlePasteIngredients}>
-                                {group.ingredients.map((ingredient: any, ingredientIndex) => (
-                                  <div key={ingredient.localId} className="p-3 border border-border/50 rounded-lg bg-background/50">
-                                    <div className="flex items-center justify-between mb-3">
-                                      <div className="flex items-center gap-2">
-                                        <Checkbox
-                                          checked={Boolean(selectedIds.has(ingredient.localId))}
-                                          onCheckedChange={(checked) =>
-                                            toggleRowSelected(ingredient.localId, Boolean(checked))
-                                          }
-                                        />
-                                        <h5 className="text-sm font-medium text-muted-foreground">
-                                          Ingredient #{ingredientIndex + 1}
-                                        </h5>
-                                      </div>
-                                      <Button
-                                        type="button"
-                                        variant="ghost"
-                                        size="sm"
-                                        onClick={() => removeIngredient(ingredientIndex)}
-                                        className="w-6 h-6 p-0 text-destructive hover:bg-destructive/10"
-                                        disabled={group.ingredients.length === 1}
-                                      >
-                                        <X className="w-3 h-3" />
-                                      </Button>
-                                    </div>
-
-                                    <div className="grid grid-cols-1 sm:grid-cols-12 gap-3">
-                                      <div className="sm:col-span-5">
-                                        <Label className="text-xs font-medium text-muted-foreground mb-1 block">Name *</Label>
-                                        <Field
-                                          as={Input}
-                                          name={`ingredientGroups[${groupIndex}].ingredients[${ingredientIndex}].name`}
-                                          placeholder="Ingredient name"
-                                          className="text-sm"
-                                        />
-                                        <ErrorMessage
-                                          name={`ingredientGroups[${groupIndex}].ingredients[${ingredientIndex}].name`}
-                                          component="p"
-                                          className="text-destructive text-xs mt-1"
-                                        />
-                                      </div>
-
-                                      <div className="sm:col-span-3">
-                                        <Label className="text-xs font-medium text-muted-foreground mb-1 block">Quantity *</Label>
-                                        <Field
-                                          as={Input}
-                                          name={`ingredientGroups[${groupIndex}].ingredients[${ingredientIndex}].quantity`}
-                                          placeholder="Amount"
-                                          type="number"
-                                          step="0.000001"
-                                          min="0"
-                                          className="text-sm"
-                                        />
-                                        <ErrorMessage
-                                          name={`ingredientGroups[${groupIndex}].ingredients[${ingredientIndex}].quantity`}
-                                          component="p"
-                                          className="text-destructive text-xs mt-1"
-                                        />
-                                      </div>
-
-                                      <div className="sm:col-span-2">
-                                        <Label className="text-xs font-medium text-muted-foreground mb-1 block">Unit *</Label>
-                                        <Field name={`ingredientGroups[${groupIndex}].ingredients[${ingredientIndex}].unit`}>
-                                          {({ field }: { field: any }) => (
-                                            <Select
-                                              value={field.value}
-                                              onValueChange={(value) =>
-                                                field.onChange({ target: { name: field.name, value } })
-                                              }
-                                            >
-                                              <SelectTrigger className="text-sm">
-                                                <SelectValue />
-                                              </SelectTrigger>
-                                              <SelectContent>
-                                                {UNIT_OPTIONS.map((option) => (
-                                                  <SelectItem key={option.value} value={option.value}>
-                                                    {option.label}
-                                                  </SelectItem>
-                                                ))}
-                                              </SelectContent>
-                                            </Select>
-                                          )}
-                                        </Field>
-                                      </div>
-
-                                      <div className="sm:col-span-2">
-                                        <Label className="text-xs font-medium text-muted-foreground mb-1 block">Cost/Unit</Label>
-                                        <div className="relative">
-                                          <DollarSign className="absolute left-2 top-1/2 transform -translate-y-1/2 w-3 h-3 text-muted-foreground" />
-                                          <Field
-                                            as={Input}
-                                            name={`ingredientGroups[${groupIndex}].ingredients[${ingredientIndex}].costPerUnit`}
-                                            placeholder="0.00"
-                                            type="number"
-                                            step="0.000001"
-                                            min="0"
-                                            className="pl-6 text-sm"
-                                          />
-                                        </div>
-                                        <ErrorMessage
-                                          name={`ingredientGroups[${groupIndex}].ingredients[${ingredientIndex}].costPerUnit`}
-                                          component="p"
-                                          className="text-destructive text-xs mt-1"
-                                        />
-                                      </div>
-                                    </div>
-                                  </div>
-                                ))}
-
-                                <Button
-                                  type="button"
-                                  variant="outline"
-                                  size="sm"
-                                  onClick={() => {
-                                    pushIngredient({
-                                      name: "",
-                                      quantity: "",
-                                      unit: DEFAULT_UNIT,
-                                      costPerUnit: "",
-                                      localId: generateStableId(),
-                                    });
-                                  }}
-                                  className="w-full border-dashed"
-                                >
-                                  <Plus className="w-3 h-3 mr-2" />
-                                  Add Ingredient to {group.name || "Group"}
-                                </Button>
-                              </div>
-                            )}
-                          </FieldArray>
-                        </div>
-                      ))}
-
-                      {/* Cost Summary */}
-                      <div className="mt-4 p-3 bg-muted/30 rounded-lg border border-border">
-                        <div className="flex items-center justify-between">
-                          <span className="text-sm font-medium text-foreground">
-                            {t("recipes.estimatedTotalCost")}:
-                          </span>
-                          <span className="text-lg font-bold text-primary">
-                            $
-                            {calculateTotalCost(
-                              values.ingredientGroups
-                            ).toFixed(2)}
-                          </span>
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
-                )}
-              </FieldArray>
+              <IngredientsInput
+                name="ingredientGroups"
+                ingredientGroups={values.ingredientGroups}
+                onFieldChange={setFieldValue}
+                selectedIds={selectedIds}
+                onSelectionChange={setSelectedIds}
+                generateStableId={generateStableId}
+                title="Ingredient Groups"
+                description="Organize ingredients into logical groups like 'Dough', 'Filling', 'Sauce'"
+                showCostSummary={true}
+                quantityType="string"
+                onPasteIngredients={handlePasteIngredients}
+              />
 
               {/* Instructions Section */}
               <Card>
